factorize code
